requirements.txt
streamlit==1.28.1
pandas
Pillow
numpy